Closes #6 - Add configuration in the main app for which county to use
- gen_nws_counties_xml.pl scrapes the NWS website and grabs the URLS for
every state, territory, district, and zone that the NWS supports and
generates the XML files for the preferences. I was originally going to
have it fetch the counties at runtime when the user picked a state, but
the tech at the NWS that I was communicating with about writing this
said they'll never change, so just hardcode them in the app.  On the off
chance they ever do change, just run this script and rebuild to pick up
the changes.