Fixes #16: Widget refreshes on KitKat now.
Did overhaul of how the threading worked on the background service,
turns out the AsyncTask wasn't necessary because we were already
running in a background thread so we didn't need to spawn a new thread
for the URL fetch.  This turned out to be the main source of the
problem (it was broken on ICS and fixing it there broke it on Kit Kat,
the way it's written now works on both).